Camera manager complete. need to add database code
