+++ RetroX_Kotlin/app/src/main/res/layout/fragment_dashboard.xml
@@ -10,19 +10,24 @@
         android:layout_height="match_parent"
         tools:context=".ui.dashboard.DashboardFragment">
 
-        <TextView
-            android:id="@+id/text_dashboard"
+        <androidx.appcompat.widget.AppCompatAutoCompleteTextView
+            android:id="@+id/discussion_search"
+            android:layout_width="0dp"
+            android:layout_height="0dp"
+            android:hint="@string/search_discussion_post"
+            app:layout_constraintHeight_percent=".05"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toTopOf="parent"/>
+
+        <androidx.recyclerview.widget.RecyclerView
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="8dp"
-            android:layout_marginTop="8dp"
-            android:layout_marginEnd="8dp"
-            android:textAlignment="center"
-            android:textSize="20sp"
+            android:layout_height="0dp"
+            app:layout_constraintTop_toBottomOf="@id/discussion_search"
+            app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
+            />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
